Complete write of SkeletonDancer (part 3)
I promised myself to avoid major rewrites like this, but spreading this over multiple
commits or pull-request would only add to my already big todo list and level of pain.

This major rewrite removes more features then it adds, the majority of features was highly unstable
or unworkable for complex use-cases. In this rewrite I tried to focus on the least minimum required
for what’s actually needed, rather then what looks cool or is nice-to-have.

The focus of this rewrite is to use templates as much as possible, while still allowing
the use of generators for more advanced use-cases.

Instead of keeping the dances (former profiles) in the project directory they now need to be installed
(globally), and cannot be customized like previously. If you need a custom dance, fork it/create it.

Note: A Dance is defined in the JSON forma, uses schema-validation and doesn’t allow importing
configuration from other dances.